backend/requirements.txt
# Django and REST framework
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers

# Database
psycopg2-binary
pgvector

# Document processing
PyPDF2
pytesseract
langchain
langchain-community
langchain-core
langchain-text-splitters
# Removed specific versions to avoid conflicts

# Vector embeddings
faiss-cpu
sentence-transformers

# Google AI integration
google-generativeai

# Task queue and caching
celery
redis   

# Utils
python-dotenv
pillow

# Testing
pytest
pytest-django